+++ src/main/java/colruyt/rearulmgtdmnejb/entity/RefRuleTypeLangPK.java
@@ -9,21 +9,21 @@ import javax.persistence.Embeddable;
 public class RefRuleTypeLangPK implements Serializable {
 	private static final long serialVersionUID = 1L;
 
-	@Column(name="RULETYPE_ID")
-	private long ruleTypeId;
-	
-	@Column(name="ISO_LANG_CODE")
+	@Column(name = "RULETYPE_ID")
+	private int ruleTypeId;
+
+	@Column(name = "ISO_LANG_CODE")
 	private String isoLangCode;
-	
-	public RefRuleTypeLangPK(){
-		
+
+	public RefRuleTypeLangPK() {
+
 	}
 
-	public long getRuleTypeId() {
+	public int getRuleTypeId() {
 		return ruleTypeId;
 	}
 
-	public void setRuleTypeId(long ruleTypeId) {
+	public void setRuleTypeId(int ruleTypeId) {
 		this.ruleTypeId = ruleTypeId;
 	}
 
@@ -36,44 +36,35 @@ public class RefRuleTypeLangPK implements Serializable {
 	}
 
 	public boolean equals(Object obj) {
-		if (this == obj)
-		{
+		if (this == obj) {
 			return true;
 		}
-		if (obj == null)
-		{
+		if (obj == null) {
 			return false;
 		}
-		if (getClass() != obj.getClass())
-		{
+		if (getClass() != obj.getClass()) {
 			return false;
 		}
 		RefRuleTypeLangPK other = (RefRuleTypeLangPK) obj;
-		if (isoLangCode == null) 
-		{
-			if (other.isoLangCode != null)
-			{
+		if (isoLangCode == null) {
+			if (other.isoLangCode != null) {
 				return false;
 			}
-		} 
-		else if (!isoLangCode.equals(other.isoLangCode))
-		{
+		} else if (!isoLangCode.equals(other.isoLangCode)) {
 			return false;
 		}
-		if (ruleTypeId != other.ruleTypeId)
-		{
+		if (ruleTypeId != other.ruleTypeId) {
 			return false;
 		}
 		return true;
 	}
-	
+
 	public int hashCode() {
 		final int prime = 31;
 		int result = 1;
 		result = prime * result + ((isoLangCode == null) ? 0 : isoLangCode.hashCode());
-		result = prime * result + (int) (ruleTypeId ^ (ruleTypeId >>> 32));
+		result = prime * result + (ruleTypeId ^ (ruleTypeId >>> 32));
 		return result;
 	}
 
-	
 }
